+++ plugin/ios/ExpoAdapterBraze/PushStories/NotificationViewController.swift
@@ -1,0 +1,5 @@
+#if canImport(BrazePushStory)
+  import BrazePushStory
+
+  class NotificationViewController: BrazePushStory.NotificationViewController {}
+#endif