+++ plugin/ios/ExpoAdapterBraze/RichPush/NotificationService.swift
@@ -1,0 +1,5 @@
+#if canImport(BrazeNotificationService)
+  import BrazeNotificationService
+
+  class NotificationService: BrazeNotificationService.NotificationService {}
+#endif